requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
starlette==0.38.6
Jinja2==3.1.6

# ─── AI / LLM ─────────────────────────────────────────────────────────────────
openai==1.51.0
anthropic==0.97.0
langchain==0.3.0
langchain-core==0.3.63
langchain-openai==0.2.0
langchain-text-splitters==0.3.8
langgraph==0.2.35
langgraph-checkpoint==2.1.2
langsmith==0.1.147
tiktoken==0.12.0

# ─── Data Validation ──────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
annotated-types==0.7.0

# ─── Database ─────────────────────────────────────────────────────────────────
supabase==2.7.4
postgrest==0.16.11
gotrue==2.12.4
realtime==2.5.3
storage3==0.7.7
supafunc==0.5.1

# ─── HTTP Clients ─────────────────────────────────────────────────────────────
httpx==0.27.2
httpcore==1.0.9
requests==2.32.5
aiohttp==3.13.3
tenacity==8.3.0

# ─── QuickBooks OAuth2 & API ──────────────────────────────────────────────────
requests-oauthlib==2.0.0
oauthlib==3.3.1
PyJWT==2.12.1

# ─── Monday.com API (GraphQL) ─────────────────────────────────────────────────
monday==2.2.1

# ─── PDF Quote Generation ─────────────────────────────────────────────────────
reportlab==4.2.5
pillow==10.4.0

# ─── Document Processing ──────────────────────────────────────────────────────
python-docx==1.2.0
openpyxl==3.1.5
PyMuPDF==1.24.10

# ─── Notifications ────────────────────────────────────────────────────────────
sendgrid==6.11.0
twilio==9.3.4

# ─── Logging & Monitoring ─────────────────────────────────────────────────────
structlog==24.4.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
orjson==3.11.7
typing_extensions==4.15.0
anyio==4.12.1
sniffio==1.3.1
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
urllib3==2.6.3
packaging==24.2
six==1.17.0
